### 2025-08-14 #8810 @editorialbot
[PRE REVIEW]: NPLinker 2: a modular and customizable framework for paired omics analyses
**Submitting author:** @justinjjvanderhooft (<a href="http://orcid.org/0000-0002-9340-5511">Justin J.J. van der Hooft</a>)
**Repository:** https://github.com/NPLinker/nplinker
**Branch with paper.md** (empty if default branch):
**Version:** v2.0.0-beta.1
