### 2025-08-15 #8810 @Kevin-Mattheus-Moerman
Note to editors, the authors declare this paper for version 1 of the software: https://doi.org/10.1371/journal.pcbi.1008920. They here seek publication of version 2.
