### 2025-09-17 812f02d
fix: add validation for unique BGC gbk file names in AntismashBGCLoader
