### 2025-10-31 1242176
add full results flag
