+++ config/nplinker.toml
@@ -80,6 +80,10 @@ parameters = "version1_parameters_or_version2_parameters"
 # BiG-SCPAPE v2 also runs a `--mix` analysis by default, so you don't need to set this parameter here.
 # An example value could be: "--mibig-version 3.1 --include-singletons --gcf-cutoffs 0.30"
 
+full_results = false
+# [REQUIRED-UNDER-CONDITIONS] Only required for BiG-SCAPE 2
+# Whether to generate a full set of results instead of only the output database for BiG-SCAPE 2
+
 [gnps]
 # Settings for GNPS.
 
