### 2025-10-31 cb0be50
add missing docstring
