### 2025-11-28 #8810 @espottesmith
@ctuni @asntech @rpetit3 @apraga @douweschulte - can any of y'all contribute your expertise to review this submission to JOSS?
