### 2025-12-19 #9528 @bgyori
The submitting author seems appropriate as a senior author in the traditional academic sense but is not a software contributor. The author list in general is appropriate and I have no major issue with this.

I find it odd that the latest stable release of NPLinker is from 2022 and version 2+ is only available as pre-release. Is there a particular reason why version 2 (the subject of this publication) cannot be released as a regular, stable release?

Otherwise the installation worked as expected, despite the fairly complex mechanism of first installing a PyPI package and then using a custom script to do further installation and downloads.

I first attempted to follow https://nplinker.github.io/nplinker/latest/quickstart/ in `local` mode.

1. I create a working directory, so far so good
2. I created an nplinker.toml file as described in the documentation.

- Given that nplinker.toml is referenced as a file in later steps and is assumed to be at the root of the working directory, is it necessary to have to explicitly define an absolute path for a root_dir in the nplinker.toml?
- The quickstart didn't make it obvious that nplinker.toml should be inside nplinker_quickstart, not right outside.

3. Preparing input data for local mode

I used the sample code for GNPS data download (task ID c22f44b14a3d450eb836d607cb9521bb) which appears to have worked.

However, I couldn't get past the AntiSMASH data step. There is no clear documentation on where to download data necessary for this step. Would it be possible to provide sample data for the purposes of the quickstart tutorial?

I tried proceeding without the AntiSMASH to see if there is some automated download process (given that I didn't find instructions on how to download it) but it errored with "FileNotFoundError: antiSMASH data directory not found"

The documentation does include a quickstart but as described above it is fairly complex and appears to be missing some descriptions necessary to work.

The paper does describe a number of tools that NPLinker integrates (or could integrate) but I didn't see any statement about whether there are other tools similar to NPLinker itself, or whether it is the first of its kind. This should be clarified in the paper.
