### 2026-01-23 80add0c
[add] more state of the art - other methods + references
