### 2026-01-23 d1887f4
[add] more state of the art - other methods + references (#343)
Co-authored-by: Rosina Torres Ortega and Justin van der Hooft <jjjvanderhooft@gmail.com>
