+++ joss/paper.bib
@@ -168,3 +168,56 @@
 	keywords = {Bioinformatics, Biosynthesis, Natural products},
 	pages = {60--68},
 }
+
+@article{medema2014pep2path,
+	title={Pep2Path: automated mass spectrometry-guided genome mining of peptidic natural products},
+	author={Medema, Marnix H and Paalvast, Yared and Nguyen, Don D and Melnik, Alexey and Dorrestein, Pieter C and Takano, Eriko and Breitling, Rainer},
+	journal={PLoS computational biology},
+	volume={10},
+	number={9},
+	pages={e1003822},
+	year={2014},
+	publisher={Public Library of Science San Francisco, USA}
+	}
+
+@article{cao2019metaminer,
+	title={MetaMiner: a scalable peptidogenomics approach for discovery of ribosomal peptide natural products with blind modifications from microbial communities},
+	author={Cao, Liu and Gurevich, Alexey and Alexander, Kelsey L and Naman, C Benjamin and Le{\~a}o, Tiago and Glukhov, Evgenia and Luzzatto-Knaan, Tal and Vargas, Fernando and Quinn, Robby and Bouslimani, Amina and others},
+	journal={Cell systems},
+	volume={9},
+	number={6},
+	pages={600--608},
+	year={2019},
+	publisher={Elsevier}
+	}
+
+@article{mohimani2017metarippquest,
+	title={MetaRiPPquest: a peptidogenomics approach for the discovery of ribosomally synthesized and post-translationally modified peptides},
+	author={Mohimani, Hosein and Gurevich, Alexey and Alexander, Kelsey L and Naman, C Benjamin and Leao, Tiago and Glukhov, Evgenia and Moss, Nathan A and Luzzatto-Knaan, Tal and Vargas, Fernando and Nothias, Louis-Felix and others},
+	journal={BioRxiv},
+	pages={227504},
+	year={2017},
+	publisher={Cold Spring Harbor Laboratory}
+	}
+
+@article{merwin2020deepripp,
+	title={DeepRiPP integrates multiomics data to automate discovery of novel ribosomally synthesized natural products},
+	author={Merwin, Nishanth J and Mousa, Walaa K and Dejong, Chris A and Skinnider, Michael A and Cannon, Michael J and Li, Haoxin and Dial, Keshav and Gunabalasingam, Mathusan and Johnston, Chad and Magarvey, Nathan A},
+	journal={Proceedings of the National Academy of Sciences},
+	volume={117},
+	number={1},
+	pages={371--380},
+	year={2020},
+	publisher={National Academy of Sciences}
+	}
+
+@article{boldt2026phylogeny,
+	title={Phylogeny-Aware Metabologenomics Accurately Assigns Natural Products to Biosynthetic Gene Clusters},
+	author={Boldt, Judith and Porten, Christoph and Haeckl, FP Jake and Hug, Joachim J and Panter, Fabian and Steglich, Matthias and Wink, Joachim and Overmann, J{\"o}rg and G{\"o}ker, Markus and Krug, Daniel and others},
+	journal={Microbial Biotechnology},
+	volume={19},
+	number={1},
+	pages={e70298},
+	year={2026},
+	publisher={Wiley Online Library}
+	}
