### 2026-01-23 e9b6a99
docs: add download instructions for example input data for local mode
