### 2026-01-23 ae823fa
docs: clarify location of nplinker.toml file and specification of root_dir in quickstart tutorial
