### 2026-01-27 d1f5917
Add httpx version constraint to prevent compatibility issues
Co-authored-by: CunliangGeng <9798985+CunliangGeng@users.noreply.github.com>
