### 2026-01-27 b3b3ca5
Add helper function to handle 404 errors in antiSMASH data downloads
