### 2026-01-27 892be14
Enhance tests with 404 error handling
