### 2026-01-27 6c25896
Replace SonarCloud action with SonarQube action
