+++ .github/workflows/sonar-cloud.yml
@@ -50,8 +50,7 @@ jobs:
           install-nplinker-deps --run-on-github
       - name: Run unit tests with coverage
         run: pytest --cov=nplinker --cov-report=xml --cov-report=term tests/unit
-      - name: SonarCloud Scan
-        uses: SonarSource/sonarcloud-github-action@master
+      - name: SonarQube Scan
+        uses: SonarSource/sonarqube-scan-action@v7.0.0 
         env:
-          GITHUB_TOKEN: ${{secrets.GITHUB_TOKEN }}
           SONAR_TOKEN: ${{secrets.SONAR_TOKEN }}
