### 2026-01-27 ea1d237
Fix formatting of SONAR_TOKEN in workflow file
