+++ .github/workflows/sonar-cloud.yml
@@ -53,4 +53,4 @@ jobs:
       - name: SonarQube Scan
         uses: SonarSource/sonarqube-scan-action@v7.0.0 
         env:
-          SONAR_TOKEN: ${{secrets.SONAR_TOKEN }}
+          SONAR_TOKEN: ${{ secrets.SONAR_TOKEN }}
