### 2026-01-28 5262a3a
Update docs/quickstart.md
Co-authored-by: Cunliang Geng <c.geng@esciencecenter.nl>
