### 2026-01-28 60013e6
Update docs/quickstart.md
Co-authored-by: Cunliang Geng <c.geng@esciencecenter.nl>
