### 2026-01-28 426aac4
move example dataset instructions
