### 2026-01-28 bef99e9
add more description to example dataset instructions
