### 2026-01-28 ae905dc
Provide example input data for the local mode (#344)
* Commit from GitHub Actions (Create JOSS PDF)

* docs: add download instructions for example input data for local mode

* docs: clarify location of nplinker.toml file and specification of root_dir in quickstart tutorial

* Update docs/quickstart.md

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* Update docs/quickstart.md

Co-authored-by: Cunliang Geng <c.geng@esciencecenter.nl>

* Update docs/quickstart.md

Co-authored-by: Cunliang Geng <c.geng@esciencecenter.nl>

* move example dataset instructions

* add more desc
[…]
