### 2026-01-28 5287105
remove `--pre` for pip install
