+++ docs/index.md
@@ -4,9 +4,3 @@
 NPLinker is a python framework for data mining microbial natural products by integrating genomics and metabolomics data.
 
 For a deep understanding of NPLinker, please refer to the [original paper](https://doi.org/10.1371/journal.pcbi.1008920).
-
-
-!!! warning "Under Development"
-    
-    NPLinker v2 is under active development ([see its pre-releases](https://pypi.org/project/nplinker/#history)). 
-    The documentation is not complete yet. If you have any questions, please contact us via [GitHub Issues](https://github.com/NPLinker/nplinker/issues).
