### 2026-01-28 85d96d8
Move installation doc to home page
