### 2026-01-28 2d599de
Update version to 2.0.0
