+++ CITATION.cff
@@ -45,7 +45,7 @@ authors:
   -
     given-names:  Marnix
     family-names:  Medema
-version: "2.0.0-beta.1"
+version: "2.0.0"
 repository-code: "https://github.com/NPLinker/nplinker"
 keywords:
   - Genome
