### 2026-01-30 #9528 @justinjjvanderhooft
Hi @espottesmith, the short answer is good (at least, we think so). I will prepare another post for the reviewers and yourself.
