### 2026-01-30 #9528 @justinjjvanderhooft
Dear @apraga @bgyori @espottesmith ,

Over the last weeks, we have worked on the various points you raised and we have done our best to either clarify ourselves in the manuscript or NPLinker instructions, or we have formulated a response to your questions. Below, you found your points collated from both reviewers per theme, with our response and any actions if they took place.

We look forward to hear back from you if additional clarifications are required or if this resolves your points,

Thanks for taking the time to review our NPLinker package,

Kind regards,

The NPLinker team

-- --

Response to reviewers

Note: the ^^text^^ are the original comments from reviewers and they are followed by our responses.

Point 1
^^A sentence about the target audience would be appreciated.^^

We added and specified information on the target audience in two places in the statement of need (see PR 340):

“Hence, natural products discovery would benefit from up-to-date and user-friendly software packages”, → “Hence, researchers working in the natural products discovery field, or anyone with paired genomics and metabolomics data, would benefit from…”

“Here, we redesigned NPLinker to provide such an integrative omics tool that guides  both users and developers in paired omics mining with its modular setup.” also specifies that we’re targeting both users and developers working with paired omics but maybe separating both would make it clearer: “Here, we redesigned NPLinker to provide such an integrative omics tool that guides users in paired omics mining with its modular setup. The tool is also of interest to developers in this field, as recent developments...”

Point 2
^^The paper does describe a number of tools that NPLinker integrates (or could integrate) but I didn't see any statement about whether there are other tools similar to NPLinker itself, or whether it is the first of its kind. This should be clarified in the paper.
&
This is not mentioned in the paper. It is not my area of expertise but NPOmix seems to have some similar features. It may be a strong selling point, so adding a sentence in the paper would improve its overall quality.^^

After the paragraph: “and metabolomics (e.g., MS/MS spectra) to elucidate gene–metabolite relationships and accelerate discovery (Goering et al., 2016; Hooft et al., 2020; Leão et al., 2022). we added new sentences and references (see PR 342 and 343):

“Several computational strategies have been developed to propose such gene cluster-mass spectral links, including i) feature-based approaches that match predicted structural or substructure information between genomes and MS/MS (e.g., Pep2Path, RiPPquest, MetaMiner, DeepRiPP) (Medema et al., 2014; Mohimani et al., 2014; Cao et al., 2019; Merwin et al., 2020), ii) correlation-based “metabologenomics” that infer co-occurrence patterns across strains or samples (Goering et al., 2016), and iii) hybrid frameworks such as NPLinker and the machine learning classifier-based NPOmix, (Eldjárn et al., 2021; Leão et al., 2022), with recently proposed phylogeny-aware extensions to reduce false positive associations (Boldt et al., 2026). However, the connected omics data structures, preproccessing pipelines, resources, and annotation tools are constantly being improved:curated BGC repositories such as MIBiG are continually expanded with new entries and annotation fields (Zdouc et al., 2025), while community MS/MS resources such as GNPS spectral libraries keep growing (Wang et al., 2016).Together with the constant expansion of experimental datasets, this puts a strain on downstream frameworks that integrate processed omics data and link results.”

Point 3
^^Installation issues on Gentoo^^

The installation issues on Gentoo have been fixed in PR 338.

Point 4
^^NPLinker versioning^^

Only pre-releases (beta versions) were released before, to install that, the --pre parameter is required, e.g., pip install –pre nplinker.  As we are moving towards publication of the tool, we now released the regular version to make the installation easier, so now pip install nplinker will work and install the version 2.0.0.

Point 5

^^Functionality: Have the functional claims of the software been confirmed?
Preparing input data for local mode
I used the sample code for GNPS data download (task ID c22f44b14a3d450eb836d607cb9521bb) which appears to have worked.
However, I couldn't get past the AntiSMASH data step. There is no clear documentation on where to download data necessary for this step. Would it be possible to provide sample data for the purposes of the quickstart tutorial?
I tried proceeding without the AntiSMASH to see if there is some automated download process (given that I didn't find instructions on how to download it) but it errored with "FileNotFoundError: antiSMASH data directory not found"

Functionality: Have the functional claims of the software been confirmed?
The documentation assume we already have data to test. A minimal test of data (real or not) woould have been nice. CI does not seem to have that either.
I could not make it work with this PODB project 9f759c90-2ffc-46e3-ac37-7317ab3f96eb.1:
ValueError: Not match PODP adapted schema, here are the detailed error:  - $.metabolomics.project: 'molecular_network' is a required property^^

Thanks for testing it.

Indeed, the “local mode” in the documentation (https://nplinker.github.io/nplinker/latest/quickstart/) assumes that target users have their own data ready to use NPLinker - this is exactly what the local mode is aiming for. Following your suggestion, we have added example data in the QuickStart part of the documentation (added in PR #344) so that any user can directly try NPLinker with these example data. Actually this example data (https://zenodo.org/records/10822604) has been automatically used by the CI, or more specifically the integration tests.

The error of the [PODB project 9f759c90-2ffc-46e3-ac37-7317ab3f96eb.1](https://pairedomicsdata.bioinformatics.nl/projects/9f759c90-2ffc-46e3-ac37-7317ab3f96eb.1) suggests that the “molecular_network” property is missing in the PODP project data which is the input of NPLinker. Molecular network data is required by NPLinker to execute its regular functionalities, so if that is missing in the input, NPLinker will return an error for that. In summary, that is expected behaviour.

Point 6
^^Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?
Not to my knowledge^^

The unit tests and integration tests are available in the codebase (https://github.com/NPLinker/nplinker/tree/dev/tests), and CI (https://github.com/NPLinker/nplinker/blob/dev/.github/workflows/build.yml) is also available to automatically run these tests when there is a new commit or PR. To manually run these tests, the [README.dev.md](http://readme.dev.md/) file (https://github.com/NPLinker/nplinker/blob/dev/README.dev.md) provides the instructions. During this project we have been following the software engineering guide of Netherlands eScience center to ensure sufficient tests and clear documentation.
Are these tests that you refer to? If not, could you please further clarify it?

Point 7
^^I created an nplinker.toml file as described in the documentation.
Given that nplinker.toml is referenced as a file in later steps and is assumed to be at the root of the working directory, is it necessary to have to explicitly define an absolute path for a root_dir in the nplinker.toml?
The quickstart didn't make it obvious that nplinker.toml should be inside nplinker_quickstart, not right outside.^^

Users can put the nplinker.toml file anywhere they like, and it does not have to stay in the root of the working directory. This design profiles a flexible way for users to manage config files and actual data files. Considering this, the root_dir in the nplinker.toml file must be explicitly specified by the user, so that the NPLinker software will know where to get and store data files.
We made several changes in the QuickStart (see PR #344) to clarify the location of nplinker.toml file and the specification of the root_dir.

Point 8
^^Has the submitting author (@justinjjvanderhooft) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
Submitting author is the project leader and has not contributed to the code. This is not an issue for me as the order of authors reflectrs the number of contributions in git. & The submitting author seems appropriate as a senior author in the traditional academic sense but is not a software contributor. The author list in general is appropriate and I have no major issue with this.^^

The assessment of the reviewers is correct - as the project formally ended during the summer of 2025, the project lead has taken the role of submitting author.
