### 2026-02-20 #9528 @espottesmith
At present, my responses aren't relevant. Ideally, I don't want to get involved until the post-review stage (where I will give feedback on paper/code, as needed). If I get involved, it's because of some major disagreement between authors and reviewers.

We've been waiting on reviewer responses for a few weeks. There's no hard and fast line, but at a certain point, if we don't hear from them, I may decide to seek out other reviewers.
