### 2026-03-03 #9528 @bgyori
The default `pip install nplinker` now installs version 2.0.0 so this was improved during the revision.

After the revision I re-ran the quickstart following the instructions and it seemed to work this time. I found one small issue in the instructions that I submitted a PR for here: https://github.com/NPLinker/nplinker/pull/346

This was improved in the revision.

This was added during the revision.
