### 2026-03-03 564771e
Change cp command to copy directories recursively (#346)
While following the quickstart for the review at https://github.com/openjournals/joss-reviews/issues/9528 I noticed this issue: `cp -r` is required here to copy subfolders like antismash/ correctly.
