### 2026-03-08 #9528 @apraga
Thanks @justinjjvanderhooft for updating the manuscript and the code. I do have 3 more nitpicks:

1. In the manuscript, `MS/MS` should be replace by `tandem mass spectrometry` in `feature-based approaches that match predicted structural or substructure information between genomes and MS/MS`.
2. The manuscript refers to configuration files in YAML while the quickstart use TOML format.
3. Regarding quickstart, I was still unable to run with pdb model. The database 3b487550-054d-4951-a830-ad7ddfcdb43c.2 does have a Molecular Network Task ID but it fails on my machine with

```
  File "/home/alex/.local/venv/nplinker/lib/python3.13/site-packages/nplinker/genomics/utils.py", line 341, in get_mappings_strain_id_bgc_id
    resolved_genome_id = mappings_original_genome_id_resolved_genome_id[original_genome_id]
                         ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^^^^^^^^^^^^^^^^^^^^
KeyError: 'PTJM01000065.1'
```
Could you provide a database ID that works in the quickstart ?

Also in the quickstart for local mode, in "Download & Extract GNPS data", `cd nplinker_quickstart` is not python code per se.
