### 2026-03-12 42dd504
Enhance run_bigscape to convert underscores to hyphens for BiG-SCAPE v2 compatibility
