### 2026-03-12 764e4d1
Update BiG-SCAPE v2 script path to remove version suffix
