### 2026-03-13 4f333b4
Enhance run_bigscape for BiG-SCAPE v2 compatibility (#349)
* Enhance run_bigscape to convert underscores to hyphens for BiG-SCAPE v2 compatibility

* Fix BiG-SCAPE v2 database file naming in DatasetArranger

* Fix formatting

* Update BiG-SCAPE installation script for compatibility and versioning improvements

* Update BiG-SCAPE v2 script path to remove version suffix

* Add importlib-resources to BiG-SCAPE installation dependencies

* Add scikit-learn to BiG-SCAPE installation dependencies
