### 2026-03-14 #9528 @justinjjvanderhooft
@apraga @espottesmith we have worked on the above points, and here are our responses/fixes:

Point 1) the text in the manuscript was updated using the reviewer's suggestion.

Point 2) thanks for spotting this, the .yaml format was previously used, we have now updated this to the .toml format that is currently used.

Point 3) We were unable to reproduce the reported KeyError in a reproducible manner, but we encountered a related error: This PoDP project contains only one genome, and it could not be fetched from antiSMASH DB. For the PoDP mode, NPLinker depends on external data providers and, unfortunately, the antiSMASH database does not contain data for all RefSeq genomes. NPLinker can only run if antiSMASH data is available for at least some genomes in a project. We also like to point out that external servers like GNPS-MassIVE can have hiccups causing the GNPS data download to be incomplete or stalled, leading to subsequent issues for NPLinker to run correctly. In such cases, rerunning NPLinker in PoDP mode later, or downloading the data separately and running NPLinker in local mode is advised as possible solutions.
We advise to use the project with ID "707bce6a-5000-43e0-95bf-ddb41cb2ccfa.3", which we have confirmed contains genomes available in the antiSMASH database to run NPLinker in PoDP mode.

Regarding the final unnumbered point on the local mode quickstart Python code: we made a small fix to transfer this into Python code.

Thanks again for your constructive review!
