### 2026-03-20 cacbb40
Add example `podp_id` to quickstart (#350)
Added example popd project id (JOSS review action point)
