### 2026-03-25 #9528 @espottesmith
As I understand it, they said that your (then-proposed) changes _would_ address all of their concerns, but they haven't commented on the changes you actually made. And they haven't updated their checklist, which we use as a way to ensure that nothing has been missed.
