### 2026-03-25 #9528 @apraga
The tool has evolved to take into account latest developments in

I had to fix the installation script to make it work on Gentoo (see my PR #338). Please note I could not run the software.
Note: the pip version is not for nplink2 but the current 1.3.2.
- [x] ~**Performance:** If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)

Partially, no minimal reproducible example though (see above).
