### 2026-03-25 #9528 @justinjjvanderhooft
Thanks @espottesmith, I will circulate a message to all authors later this week and update you here once the dots and i's on our end are crossed. One question/request: the software is being archived on Zenodo (https://zenodo.org/records/19019159), including the latest version 2.0.1; however, the author list is not identical due to historical reasons. Can we use this archive as I would prefer not to create yet another one. The license corresponds with the GitHub licence. Thanks for your consideration.
