### 2026-03-25 #9528 @espottesmith
The authors should be identical between the archive and the paper. Can you explain why the authors on the archive can't be updated to match the paper? If it's sensitive for some reason, you can e-mail me (espottesmith@gmail.com or evan.spotte-smith@ucd.ie).
