### 2026-03-30 a1a7bf3
update title and author list in CITATION.cff
