+++ pyproject.toml
@@ -6,7 +6,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "nplinker"
-version = "2.0.1"
+version = "2.0.2"
 description = "Natural Products Linker"
 readme = "README.md"
 requires-python = ">=3.11"
