### 2026-03-31 e9e5e08
Update CITATION.cff
Updated Dora's affliction and Wageningen affiliation.
