### 2026-03-31 73a46c3
Update CITATION.cff
Added SA affiliation to Justin, harmonised affiliations (city + country lettercode)
