### 2026-03-31 1e6c677
Update paper.md
Added Tubingen affiliation Dora
