### 2026-04-01 #9528 @justinjjvanderhooft
@espottesmith thanks for your additional guidance by email - I confirm that we finished with the Post-Review Checklist for Editor and Authors - here is the requested link to the (Zenodo) archive: https://zenodo.org/records/19367010 (DOI: https://doi.org/10.5281/zenodo.4680218)
