### 2026-04-02 #9528 @espottesmith
Please make sure that all author affiliations are identical between the paper and the archive. Most look good, but Dora Ferreira and @justinjjvanderhooft seem to be missing affiliations
