### 2026-04-02 #9528 @espottesmith
Your manuscript appears to (partially) repeat a paragraph (starting "Omics datasets"). I suggest rewriting to avoid this repetition.

It would be helpful for readers (especially those looking to become users of NPLinker2) to have a sense of what these external, non-PyPI dependencies are.

Text in Figure 2 is quite small and difficult to read. Please make sure all figures are as accessible as possible.

Otherwise, the manuscript is looking pretty good! We're almost at the finish line.
