+++ joss/paper.md
@@ -108,7 +108,7 @@ Modularity and extensibility are key features of NPLinker 2, which provides a se
 **"Scoring" component:** This component handles the linking of data and the scoring of those links. A undirected graph is used to store the linked data, with nodes corresponding to genomics or metabolomics data items and edges representing the links between them with scoring values, as illustrated in \autoref{fig:2}. The `ScoringBase` interface is provided to allow the implementation of custom scoring methods.
 
 
-![Graph representation of linkings. \label{fig:2}](fig2.png){ width=40%}
+![Graph representation of linkings. \label{fig:2}](fig2.png){ width=80%}
 
 ## New documentation website
 A dedicated documentation website is available to help users and developers understand how to use and extend NPLinker 2. It includes tutorials, conceptual overviews, pipeline diagrams, and an API reference. The documentation is available at [https://nplinker.github.io/nplinker](https://nplinker.github.io/nplinker).
