### 2026-04-02 b59c4f7
Update paper.md
Removed unintended duplications of text in Statement of Need section.
