### 2026-04-02 #9528 @espottesmith
Paper looks good, and I appreciate your fast work.

Please let me know when you've confirmed that you're not missing any DOIs (see above) and when you've added the missing affiliations to the Zenodo archive.
