### 2026-04-02 374a267
Update paper.bib
Added missing DOIs
