+++ joss/paper.bib
@@ -173,6 +173,7 @@
 	title={Pep2Path: automated mass spectrometry-guided genome mining of peptidic natural products},
 	author={Medema, Marnix H and Paalvast, Yared and Nguyen, Don D and Melnik, Alexey and Dorrestein, Pieter C and Takano, Eriko and Breitling, Rainer},
 	journal={PLoS computational biology},
+	doi = {10.1371/journal.pcbi.1003822},
 	volume={10},
 	number={9},
 	pages={e1003822},
@@ -184,6 +185,7 @@
 	title={MetaMiner: a scalable peptidogenomics approach for discovery of ribosomal peptide natural products with blind modifications from microbial communities},
 	author={Cao, Liu and Gurevich, Alexey and Alexander, Kelsey L and Naman, C Benjamin and Le{\~a}o, Tiago and Glukhov, Evgenia and Luzzatto-Knaan, Tal and Vargas, Fernando and Quinn, Robby and Bouslimani, Amina and others},
 	journal={Cell systems},
+	doi={10.1016/j.cels.2019.09.004},
 	volume={9},
 	number={6},
 	pages={600--608},
@@ -195,6 +197,7 @@
 	title={MetaRiPPquest: a peptidogenomics approach for the discovery of ribosomally synthesized and post-translationally modified peptides},
 	author={Mohimani, Hosein and Gurevich, Alexey and Alexander, Kelsey L and Naman, C Benjamin and Leao, Tiago and Glukhov, Evgenia and Moss, Nathan A and Luzzatto-Knaan, Tal and Vargas, Fernando and Nothias, Louis-Felix and others},
 	journal={BioRxiv},
+	doi={doi.org/10.1101/227504},
 	pages={227504},
 	year={2017},
 	publisher={Cold Spring Harbor Laboratory}
@@ -204,6 +207,7 @@
 	title={DeepRiPP integrates multiomics data to automate discovery of novel ribosomally synthesized natural products},
 	author={Merwin, Nishanth J and Mousa, Walaa K and Dejong, Chris A and Skinnider, Michael A and Cannon, Michael J and Li, Haoxin and Dial, Keshav and Gunabalasingam, Mathusan and Johnston, Chad and Magarvey, Nathan A},
 	journal={Proceedings of the National Academy of Sciences},
+	doi{10.1073/pnas.1901493116},
 	volume={117},
 	number={1},
 	pages={371--380},
@@ -215,6 +219,7 @@
 	title={Phylogeny-Aware Metabologenomics Accurately Assigns Natural Products to Biosynthetic Gene Clusters},
 	author={Boldt, Judith and Porten, Christoph and Haeckl, FP Jake and Hug, Joachim J and Panter, Fabian and Steglich, Matthias and Wink, Joachim and Overmann, J{\"o}rg and G{\"o}ker, Markus and Krug, Daniel and others},
 	journal={Microbial Biotechnology},
+	doi={10.1111/1751-7915.70298},
 	volume={19},
 	number={1},
 	pages={e70298},
