### 2026-04-02 #9528 @justinjjvanderhooft
@espottesmith I have just added the missing DOIs to the bib file.
