### 2026-04-02 #9528 @justinjjvanderhooft
@espottesmith regarding the missing affiliations in Zenodo, for me, they are the same, I have checked again the CFF file, and this matches the paper. I am attaching a screenshot of the Zenodo archive where you can see the same two affilitations for me as in the manuscript (the same is true for Dora). So I am unsure what is missing?
