### 2026-04-02 61d52ec
Update paper.bib
Fixed syntax error in bib file
