+++ joss/paper.bib
@@ -207,7 +207,7 @@
 	title={DeepRiPP integrates multiomics data to automate discovery of novel ribosomally synthesized natural products},
 	author={Merwin, Nishanth J and Mousa, Walaa K and Dejong, Chris A and Skinnider, Michael A and Cannon, Michael J and Li, Haoxin and Dial, Keshav and Gunabalasingam, Mathusan and Johnston, Chad and Magarvey, Nathan A},
 	journal={Proceedings of the National Academy of Sciences},
-	doi{10.1073/pnas.1901493116},
+	doi={10.1073/pnas.1901493116},
 	volume={117},
 	number={1},
 	pages={371--380},
