### 2026-04-06 #9528 @espottesmith
Hey @justinjjvanderhooft please fix the improperly formatted DOI
