### 2026-04-06 0d7ac2c
Update paper.bib
Fixed invalid DOI
