+++ joss/paper.bib
@@ -197,7 +197,7 @@
 	title={MetaRiPPquest: a peptidogenomics approach for the discovery of ribosomally synthesized and post-translationally modified peptides},
 	author={Mohimani, Hosein and Gurevich, Alexey and Alexander, Kelsey L and Naman, C Benjamin and Leao, Tiago and Glukhov, Evgenia and Moss, Nathan A and Luzzatto-Knaan, Tal and Vargas, Fernando and Nothias, Louis-Felix and others},
 	journal={BioRxiv},
-	doi={doi.org/10.1101/227504},
+	doi={10.1101/227504},
 	pages={227504},
 	year={2017},
 	publisher={Cold Spring Harbor Laboratory}
