### 2026-04-06 #9528 @justinjjvanderhooft
The proof pdf looks good to me 😎
