### 2026-04-27 #9528 @justinjjvanderhooft
Good day @espottesmith, any update on the status of our manuscript? Have a nice week ahead! Best, Justin
