### 2026-05-02 #9528 @arfon
@justinjjvanderhooft – please fix these issues in your manuscript:

**Typos/spelling:**

- preproccessing → preprocessing (Statement of need)
- .e.g., → e.g., (Statement of need — stray leading period)
- vice vera → vice versa (User-friendly webapp section)

**Grammar:**

- "A undirected graph" → "An undirected graph" (Scoring component)

**Duplicate word:**

- "and and UKRI EP/X03142X/2" — "and" is repeated (Acknowledgements)

**Section heading:**

- "Reference" → "References"
