### 2026-05-02 044f553
Final spelling and grammar fixes in paper.md
