### 2026-05-03 #9528 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
